+++ src/main/java/net/wizards/client/effect/FrozenParticles.java
@@ -3,6 +3,7 @@ package net.wizards.client.effect;
 import net.minecraft.entity.LivingEntity;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
 import net.spell_engine.api.spell.fx.ParticleBatch;
+import net.spell_engine.client.util.Color;
 import net.spell_engine.fx.ParticleHelper;
 import net.spell_engine.fx.SpellEngineParticles;
 
@@ -12,10 +13,9 @@ public class FrozenParticles implements CustomParticleStatusEffect.Spawner {
 
     public FrozenParticles(int particleCount) {
         this.particles = new ParticleBatch(
-                SpellEngineParticles.getMagicParticleVariant(
-                        SpellEngineParticles.FROST,
-                        SpellEngineParticles.MagicParticleFamily.Shape.IMPACT,
-                        SpellEngineParticles.MagicParticleFamily.Motion.BURST
+                SpellEngineParticles.MagicParticles.get(
+                        SpellEngineParticles.MagicParticles.Shape.FROST,
+                        SpellEngineParticles.MagicParticles.Motion.BURST
                 ).id().toString(),
                 ParticleBatch.Shape.SPHERE,
                 ParticleBatch.Origin.CENTER,
@@ -23,7 +23,8 @@ public class FrozenParticles implements CustomParticleStatusEffect.Spawner {
                 particleCount,
                 0.1F,
                 0.3F,
-                0);
+                0)
+                .color(Color.FROST.toRGBA());
     }
 
     @Override